requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
langchain==0.3.9
langgraph==0.2.55
langchain-community==0.3.9
langchain-ollama==0.2.1
langchain-core==0.3.21
qdrant-client==1.12.1
motor==3.6.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18
pypdf==5.1.0
docx2txt==0.8
pandas==2.2.3
tavily-python==0.5.0
sentence-transformers==3.3.1
PyYAML==6.0.2
httpx==0.27.2
